Move harfbuzz shaper to a new shapers module
