+++ components/fonts/shapers/harfbuzz.rs
@@ -29,6 +29,7 @@ use harfbuzz_sys::{
 use log::debug;
 use num_traits::Zero;
 
+use super::unicode_script_to_iso15924_tag;
 use crate::font::advance_for_shaped_glyph;
 use crate::platform::font::FontTable;
 use crate::{
@@ -216,18 +217,6 @@ impl Shaper {
     fn fixed_to_float(i: hb_position_t) -> f64 {
         fixed_to_float(16, i)
     }
-}
-
-pub fn unicode_script_to_iso15924_tag(script: unicode_script::Script) -> u32 {
-    let bytes: [u8; 4] = match script {
-        unicode_script::Script::Unknown => *b"Zzzz",
-        _ => {
-            let short_name = script.short_name();
-            short_name.as_bytes().try_into().unwrap()
-        },
-    };
-
-    u32::from_be_bytes(bytes)
 }
 
 impl Shaper {
